requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
imgkit==1.0.2
Pillow==6.1.0
pymaging==0.1
pymaging-png==0.1
pypng==0.0.20
pyshorteners==0.6.1
qrcode==6.1
requests==2.22.0
six==1.12.0
text-to-image==0.0.5
urllib3==1.25.3
webkit2png==0.8.2
